# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: BBLBB API
  version: 1.0.0
  description: |
    BBLBB 稳定业务 API。字段、错误、权限和安全语义以本文件及其引用的冻结文档为准。
    生产实现必须保持 SQLite、MySQL 8 和 MariaDB 10.11 的契约一致性。
  license:
    name: Proprietary
servers:
- url: https://{host}
  variables:
    host:
      default: forum.example.com
security: []
...
paths:
...
  "/api/v1/attachments/{id}/download-policy":
    get:
      tags:
      - Attachments
      operationId: get_attachments_id_download_policy
      security: *2
      x-permission: authenticated
      x-csrf: false
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '204':
          description: No content
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '429':
          "$ref": "#/components/responses/RateLimited"
      summary: get api v1 attachments id download policy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  "/api/v1/attachments/{id}/download":
    post:
      tags:
      - Attachments
      operationId: post_attachments_id_download
      security:
      - sessionCookie: []
      x-permission: authenticated
      x-csrf: true
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DownloadRequest"
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '204':
          headers:
            Cache-Control:
              description: private, no-store
              schema:
                type: string
                example: private, no-store
          description: No content
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '429':
          "$ref": "#/components/responses/RateLimited"
      summary: post api v1 attachments id download
  "/api/v1/download-authorizations/{id}":
    get:
      tags:
      - Attachments
      operationId: get_download_authorizations_id_
      security: *2
      x-permission: authenticated
      x-csrf: false
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '204':
          description: No content
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '429':
          "$ref": "#/components/responses/RateLimited"
      summary: get api v1 download authorizations id
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  "/api/v1/download-authorizations/{id}/sign-url":
    post:
      tags:
      - Attachments
      operationId: post_download_authorizations_id_sign_url
      security:
      - sessionCookie: []
      x-permission: authenticated
      x-csrf: true
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GenericRequest"
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '204':
          headers:
            Cache-Control:
              description: private, no-store
              schema:
                type: string
                example: private, no-store
          description: No content
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '429':
          "$ref": "#/components/responses/RateLimited"
      summary: post api v1 download authorizations id sign url
  "/api/v1/me/download-transactions":
    get:
      tags:
      - Attachments
      operationId: get_me_download_transactions
      security: *2
      x-permission: authenticated
      x-csrf: false
      responses:
        '200':
          "$ref": "#/components/responses/GenericSuccess"
        '204':
          description: No content
        '400':
          "$ref": "#/components/responses/ProblemResponse"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '401':
          "$ref": "#/components/responses/AuthenticationRequired"
        '429':
          "$ref": "#/components/responses/RateLimited"
      summary: get api v1 me download transactions
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bblbb_session
      description: Browser session cookie. State-changing requests also require X-CSRF-Token.
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
...
  responses:
    AuthenticationRequired:
      description: Authentication is required
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    RateLimited:
      description: Request rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
...
    GenericSuccess:
      description: Successful operation
      headers:
        Cache-Control:
          description: private, no-store
          schema:
            type: string
            example: private, no-store
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/GenericSuccess"
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
  schemas:
    GenericRequest:
      type: object
      additionalProperties: true
...
    GenericSuccess:
      type: object
      additionalProperties: true
...
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      - request_id
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          enum:
          - invalid_request
          - visibility_level_exceeds_author
          - invalid_url
          - idempotency_conflict
          - version_conflict
          - unauthorized
          - invalid_credentials
          - mfa_code_invalid
          - mfa_challenge_invalid
          - mfa_confirm_invalid
          - mfa_enrollment_invalid
          - passkey_not_configured
          - passkey_unavailable
          - passkey_challenge_invalid
          - passkey_registration_invalid
          - passkey_limit_reached
          - reauth_password_invalid
          - verification_token_invalid
          - reset_token_invalid
          - validation_failed
          - invalid_current_password
          - forbidden
          - step_up_required
          - not_found
          - csrf_failed
          - origin_not_allowed
          - host_not_allowed
          - rate_limited
          - crawler_denied
          - challenge_required
          - temporarily_banned
          - feature_disabled
          - insufficient_funds
          - daily_limit_exceeded
          - checkout_interaction_invalid
          - checkout_user_mismatch
          - checkout_intent_expired
          - checkout_intent_consumed
          - offer_version_changed
          - refund_not_allowed
          - product_unavailable
          - shop_purchase_limit_exceeded
          - shop_stock_exhausted
          - entitlement_not_usable
          - presentation_slot_conflict
          - activity_already_claimed
          - activity_not_eligible
          - invalid_price_coin
          - post_not_paid
          - price_not_configured
          - self_reaction
          - cannot_follow_self
          - cannot_message_self
          - download_url_unavailable
          - ai_consent_required
          - ai_budget_exceeded
          - ai_suggestion_stale
          - invalid_storage_request
          - storage_partial_upload
          - storage_forbidden
          - storage_auth_failed
          - storage_rate_limited
          - quota_exceeded
          - storage_conflict
          - storage_state_error
          - storage_verification_failed
          - storage_network_error
          - storage_upstream_error
          - theme_invalid
          - theme_incompatible
          - theme_not_found
          - theme_conflict
          - plugin_invalid
          - plugin_incompatible
          - plugin_not_found
          - plugin_conflict
          - marketplace_disabled
          - marketplace_invalid_client
          - refund_exceeds_purchase
          - merchant_balance_insufficient
          - webhook_invalid_signature
          - bad_request
          - conflict
          - video_insecure_scheme
          - video_invalid_url
          - video_host_invalid
          - video_port_not_allowed
          - video_private_ip
          - video_signed_url
          - video_userinfo_not_allowed
          - video_fragment_not_allowed
          - video_unsupported_type
          - video_not_video_page
          - video_invalid
          - video_mime_mismatch
          - video_no_embed_permission
          - video_takedown
          - video_provider_disabled
          - video_provider_host_not_allowed
          - video_provider_ratelimited
          - video_provider_unavailable
          - video_policy_changed
          - video_policy_version_conflict
          - video_poster_attachment_invalid
          - video_resolution_expired
          - video_embed_not_found
          - video_embed_referenced
          - video_target_conflict
          - video_target_forbidden
          - video_target_not_found
          - video_version_conflict
          - video_egress_http_error
          - video_egress_private_ip
          - video_egress_timeout
          - video_egress_too_large
          - video_egress_too_many_redirects
          - video_egress_unavailable
          - video_hls_invalid
          - video_hls_depth_exceeded
          - video_hls_segment_count_exceeded
          - video_hls_duration_exceeded
          - video_hls_cross_origin_segment
          - video_hls_key_not_allowed
          - video_hls_map_not_allowed
          - video_hls_signed_uri
          - internal_error
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            type: object
            required:
            - code
            - message_key
            properties:
              field:
                type: string
              code:
                type: string
              message_key:
                type: string
...
    DownloadRequest:
      type: object
      additionalProperties: false
      required:
      - client_request_id
      properties:
        target_type:
          type:
          - string
          - 'null'
          enum:
          - post
          - comment
        target_id:
          type:
          - string
          - 'null'
          format: uuid
        expected_policy_version:
          type:
          - integer
          - 'null'
          format: int64
        client_request_id:
          type: string
          minLength: 16
          maxLength: 200
...